+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-
+#### Fixed
+  - [#447](https://github.com/Day8/re-frame/issues/447) After interceptor runs against coeffect db if effect db is nil/false. 
 
 ## 0.10.5 (2018.02.13)